portfolio/mediflow/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
scipy>=1.12.0
numpy>=1.26.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
